+++ src/components/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/named */
+/* eslint-disable import/no-cycle */
+
+// Shared components
+export * from './marginals';
+export * from './shared';
