+++ src/components/marginals/NavBar/DesktopNav/index.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import { Link } from 'gatsby';
+import navList from '../../../../../config/content';
+import { NavBar, Logo, NavItem } from './style';
+import { MLHbanner } from '../../../shared';
+
+function DesktopNav() {
+  return (
+    <NavBar>
+      <div className='navWrapper'>
+        <div className='nav-left'>
+          <Logo src={navList.logo.src} alt={navList.logo.alt} />
+        </div>
+
+        <nav>
+          <ul className='navList'>
+            {navList.navItems.map((elem) => (
+              <li
+                key={elem.id}
+                id={elem.name}
+                role='menuitem'
+                onClick={null}
+                onKeyPress={null}
+                className='NavListItem'
+              >
+                <Link style={{ textDecoration: 'none' }} to='/' tabIndex='0'>
+                  <NavItem>{elem.name}</NavItem>
+                </Link>
+              </li>
+            ))}
+          </ul>
+        </nav>
+
+        <div className='nav-right'>
+          <MLHbanner />
+        </div>
+      </div>
+    </NavBar>
+  );
+}
+
+export default DesktopNav;
