+++ src/components/marginals/NavBar/MobileNav/index.jsx
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
-import Container from '../../atoms/Container';
-import navList from '../../../../config/content';
+import navList from '../../../../../config/content';
 import { MenuButton, NavBar, Logo } from './style';
-import MLHbanner from '../../atoms/MLHbanner';
+import {MLHbanner, Container} from '../../../shared';
 import MobileNavList from './MobileNavList/MobileNavList';
 
 function MobileNav() {
