+++ src/components/marginals/index.js
@@ -1,0 +1,3 @@
+/* eslint-disable import/no-cycle */
+export { default as DesktopNav } from './NavBar/DesktopNav';
+export { default as MobileNav } from './NavBar/MobileNav';
