+++ src/components/shared/index.js
@@ -8,3 +8,7 @@ export { default as ButtonMeta } from './Typography/ButtonMeta';
 export { default as ButtonMetaLarge } from './Typography/ButtonMetaLarge';
 export { default as Caption } from './Typography/Caption';
 export { default as NavText } from './Typography/NavText';
+
+export { default as MLHbanner } from './MLHbanner';
+export { default as Layout } from './Layout';
+export { default as Container } from './Container';
