+++ src/pages/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '../components/molecules/Layout/Layout';
+import {Layout} from '../components/shared/index';
 import '../styles/global.css';
 
 export default function Home() {
